restructured profile module
