+++ src/app/profile/change-password-dialog/change-password-dialog.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { FormControl, FormGroup } from '@angular/forms';
-import { equalValidation } from '../shared/validation';
+import { equalValidation } from '../../shared/validation';
 import { ChangePasswordData } from './change-password-data';
 
 @Component({
