+++ src/app/profile/profile/activity-item/activity-item.component.scss
