+++ src/app/profile/profile/observation-item/observation-item.component.scss
