+++ src/app/profile/profile/profile-public/profile-public.component.scss
